Widen the manor money arithmetic before it multiplies, not after
Three places in CastleManorManager multiply a price by an amount and hand the
result to something that takes a long. SeedProduction declares both as int, and
CropProcure inherits them, so the multiply happened in int arithmetic and only
the wrapped result was widened.

The one that costs money is:

	castle.addToTreasuryNoTax(crop.getAmount() * crop.getPrice());

addToTreasuryNoTax reads a negative amount as a withdrawal -- it negates it and
subtracts from the treasury. So an overflow there does not merely lose the
deposit, it takes the same sum back out. The other two feed getManorCost's long
accumulator, which is declared long precisely because the total is expected to
be large, and then loses the widening on each individual term.

Worked out how far the shipped data is from the edge rather than guessing.
RequestSetCrop validates a submitted price against the seed's own
getCropMaxPrice, which is the crop item's reference price times ten, and the
amount against getCropLimit, which is limit_crops times RateDropManor. Across
the 198 rows of Seeds.xml the largest limit_crops is 9000, and across the 54
crop items the largest reference price is 2750. At the shipped RateDropManor of
1 the worst case product is 247,500,000, comfortably inside int.

But RateDropManor scales the limit, and the threshold is
2147483647 / (9000 * 27500), which is 8.67. A server running RateDropManor at 9
or above can overflow this, and high rate servers routinely run manor rates well
past that. All three now cast before multiplying.

Noted, not changed: RequestBuySeed accumulates into an int totalPrice and checks
it against MAX_ADENA after each addition. Each term is individually bounded by an
explicit overflow pre-check above it, but two terms near the 2,000,000,000 cap
sum past Integer.MAX_VALUE and wrap negative, which passes the check rather than
failing it. It costs nothing, because the per-item reduceAdena in the purchase
loop charges the correct positive price and skips what the player cannot afford
-- only the anti-cheat punishment is evaded, and the manor prices needed to reach
it are far above what the data allows.

+++ java/org/l2jmobius/gameserver/managers/CastleManorManager.java
@@ -329,7 +329,11 @@ public class CastleManorManager implements IXmlReader
 							// Reserved and not used money giving back to treasury
 							if (crop.getAmount() > 0)
 							{
-								castle.addToTreasuryNoTax(crop.getAmount() * crop.getPrice());
+								// Widened before the multiply, not after. Both getters are int, so the product
+								// wrapped in int arithmetic and only then reached this long parameter -- and
+								// addToTreasuryNoTax reads a negative amount as a withdrawal, so an overflow
+								// here took money out of the treasury instead of putting it back.
+								castle.addToTreasuryNoTax((long) crop.getAmount() * crop.getPrice());
 							}
 						}
 					}
@@ -606,12 +610,13 @@ public class CastleManorManager implements IXmlReader
 		for (SeedProduction seed : production)
 		{
 			final Seed s = getSeed(seed.getId());
-			total += (s == null) ? 1 : (s.getSeedReferencePrice() * seed.getStartAmount());
+			total += (s == null) ? 1 : ((long) s.getSeedReferencePrice() * seed.getStartAmount());
 		}
 		
 		for (CropProcure crop : procure)
 		{
-			total += crop.getPrice() * crop.getStartAmount();
+			// Same widening as above: total is a long, but the product was computed in int first.
+			total += (long) crop.getPrice() * crop.getStartAmount();
 		}
 		
 		return total;
